Support for Traces, multiple API keys, Logger.
Appygram4J now support sending Traces to Appygram, the
preferred way for sending exceptions.  The README has
been updated with examples on how to use this.

Also, Appygram.send is now Appygram.Global.send.  This
is to allow for users to configure multiple instances of
AppygramMessenger, which are connected to differing API
keys.  Configuration works the same way; information is
in the README.

Finally, all logging is being routed through AppygramLogger.
There isn't much to log, aside from configuration errors
when trying to send an Appygram, but it's there.

See #2